OD-10915 Update FormatUtils to pass unit tests since Money is BigDecimal not Number and toString() is overridden and returns not decimal value.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49107 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-services/src/main/java/ish/oncourse/util/tapestry/TapestryFormatUtils.java
@@ -98,7 +98,7 @@ public class TapestryFormatUtils {
         String formatted = StringUtils.EMPTY;
         try {
             NumberFormat format = new DecimalFormat(pattern);
-            formatted = format.format(money);
+            formatted = format.format(money.toBigDecimal());
         } catch (Exception ex) {
             logger.error(
                     String.format("in Tapestry component during money formatting (money: %s; pattern: %s).", money, pattern),
